+++ setup/PerfTimer.h
@@ -15,7 +15,7 @@ public:
 	void unpause();
 
 	//Gets the timer's time
-	Uint32 getTicks();
+	Uint64 getTicks();
 
 	//Checks the status of the timer
 	bool isStarted();
